remove decimal

+++ catalog/app/core/product.py
@@ -1,10 +1,9 @@
 from dataclasses import dataclass
-from decimal import Decimal
 
 
 @dataclass
 class Product:
     id: int
     description: str
-    price: Decimal
+    price: float
     vendor_id: int
